minor structure fix
