+++ src/components/app/details/cdDetails/deploymentHistoryDiff/DeploymentHistoryHeader.tsx
@@ -6,7 +6,7 @@ import moment from 'moment'
 import { DEPLOYMENT_HISTORY_CONFIGURATION_LIST_MAP, Moment12HourFormat, URLS } from '../../../../../config'
 import { ReactComponent as LeftIcon } from '../../../../../assets/icons/ic-arrow-forward.svg'
 import { CompareWithBaseConfiguration, DeploymentHistoryParamsType, DeploymentTemplateOptions } from '../cd.type'
-import { Option, styles } from '../cd.utils'
+import { Option, styles } from './utils'
 import { getDeploymentDiffSelector } from '../service'
 import { showError } from '../../../../common'
 import Tippy from '@tippyjs/react'
