+++ src/components/app/details/cicdHistory/History.components.tsx
@@ -6,11 +6,12 @@ import { ReactComponent as ZoomIn } from '../../../../assets/icons/ic-fullscreen
 import { ReactComponent as ZoomOut } from '../../../../assets/icons/ic-exit-fullscreen.svg'
 import { ReactComponent as DropDownIcon } from '../../../../assets/icons/ic-chevron-down.svg'
 import { ReactComponent as OpenInNew } from '../../../../assets/icons/ic-open-in-new.svg'
+import { ReactComponent as Info } from '../../../../assets/icons/info-filled.svg'
+import { ReactComponent as Question } from '../../../../assets/icons/ic-help.svg'
 import AppNotDeployed from '../../../../assets/img/app-not-deployed.png'
 import { CiMaterial, GitTriggers, History } from './types'
-import { EVENT_STREAM_EVENTS_MAP, Host, LOGS_RETRY_COUNT, POD_STATUS, Routes } from '../../../../config'
+import { DOCUMENTATION, EVENT_STREAM_EVENTS_MAP, Host, LOGS_RETRY_COUNT, POD_STATUS, Routes } from '../../../../config'
 import { default as AnsiUp } from 'ansi_up'
-import { renderConfigurationError } from '../cdDetails/cd.utils'
 import { STAGE_TYPE } from '../triggerView/types'
 import GitCommitInfoGeneric from '../../../common/GitCommitInfoGeneric'
 import EmptyState from '../../../EmptyState/EmptyState'
@@ -130,6 +131,42 @@ function useCIEventSource(url: string, maxLength?: number) {
     }, [url, maxLength])
 
     return [data, eventSourceRef.current, logsNotAvailableError]
+}
+
+const renderLogsNotAvailable = (subtitle?: string): JSX.Element => {
+  return (
+      <div className="flexbox dc__content-center flex-align-center dc__height-inherit">
+          <div>
+          <div className="text-center"><Info className="icon-dim-20"/></div>
+          <div className="text-center cn-0 fs-14 fw-6">Logs not available</div>
+          <div className="text-center cn-0 fs-13 fw-4">{subtitle || 'Blob storage was not configured at pipeline run.'}</div>
+          </div>
+      </div>
+  )
+}
+
+const renderBlobNotConfigured = (): JSX.Element => {
+  return (
+      <>
+          {renderLogsNotAvailable('Logs are available only at runtime.')}
+          <div className="flexbox configure-blob-container pt-8 pr-12 pb-8 pl-12 bcv-1 br-4">
+              <Question className="icon-dim-20 fcv-5" />
+              <span className="fs-13 fw-4 mr-8 ml-8">Want to store logs to view later?</span>
+              <a className="fs-13 fw-6 cb-5 dc__no-decor" href={DOCUMENTATION.BLOB_STORAGE} target="_blank">
+                  Configure blob storage
+              </a>
+              <OpenInNew className="icon-dim-20 ml-8" />
+          </div>
+      </>
+  )
+}
+
+const renderConfigurationError = (isBlobStorageConfigured: boolean): JSX.Element => {
+  return (
+      <div className="flexbox dc__content-center flex-align-center dc__height-inherit">
+          {!isBlobStorageConfigured ? renderBlobNotConfigured() : renderLogsNotAvailable()}
+      </div>
+  )
 }
 
 export const LogsRenderer = ({
